BPR-68: refactor namespaces
