+++ BPR.Models/Analysis/ResultModel.cs
@@ -1,3 +1,3 @@
-﻿namespace BPRBE.Models;
+﻿namespace BPR.Models.Analysis;
 
 public record ResultModel(double Score, IEnumerable<ViolationModel> Violations);
