+++ BPR.Models/Analysis/ViolationModel.cs
@@ -1,5 +1,17 @@
-﻿using BPRBE.Enums;
+﻿using BPR.Models.Enums;
 
-namespace BPRBE.Models;
+namespace BPR.Models.Analysis;
 
-public record ViolationModel(int Id, string Name, ViolationType Type, string Description, ViolationSeverity Severity, string Code);
+// public record ViolationModel(int Id, string Name, ViolationType Type, string Description, ViolationSeverity Severity, string Code) {
+//    
+// }
+
+public class ViolationModel {
+    public int Id { get; set; }
+    public string Name { get; set; } = null!;
+
+    public ViolationType Type { get; set; }
+    public string Description { get; set; } = null!;
+    public ViolationSeverity Severity { get; set; }
+    public string Code { get; set; } = null!;
+}
