+++ BPR.Models/Persistence/Result.cs
@@ -1,4 +1,4 @@
-﻿namespace BPRBE.Models.Persistence;
+﻿namespace BPR.Models.Persistence;
 
 public class Result
 {
