+++ BPRBE.Tests/ValidatorTests.cs
@@ -1,4 +1,5 @@
-﻿using BPRBE.Models.Persistence;
+﻿using BPR.Models.Persistence;
+using BPRBE.Models.Persistence;
 using BPRBE.Validators;
 using FluentValidation;
 using FluentValidation.TestHelper;
