+++ BPRBlazor/Pages/Index.razor.cs
@@ -1,4 +1,4 @@
-﻿using BPRBlazor.Models;
+﻿using BPR.Models.Blazor;
 using Microsoft.AspNetCore.Components;
 using Microsoft.AspNetCore.Components.Forms;
 using SevenZipExtractor;
