+++ BPRBlazor/Pages/Results.razor.cs
@@ -1,4 +1,5 @@
-﻿using BPRBE.Enums;
+﻿using BPR.Models.Analysis;
+using BPR.Models.Enums;
 using BPRBE.Models;
 using Microsoft.AspNetCore.Components;
 
@@ -29,9 +30,9 @@ public partial class Results : ComponentBase
     {
         return new List<ViolationModel>
         {
-            new(1, "Violation name 1", ViolationType.ForbiddenDependencyDirection, "A class from the 'Models' namespace must not depend on a class in the 'Controllers' namespace", ViolationSeverity.Major, "using Controllers;\n\nnamespace Models;\n\npublic class One {\n    private IDependency _dependency;\n    \n    public One(IDependency dependency) {\n        _dependency = dependency;\n    }"),
-            new(2, "Violation name 2", ViolationType.Unknown, "Some other thing I have no idea...", ViolationSeverity.Minor, "public class Two {\n    private IDependency _dependency;\n    \n    public One(IDependency dependency) {\n        _dependency = dependency;\n    }"),
-            new(3, "Violation name 3", ViolationType.Unknown, "Some other thing I have no idea...", ViolationSeverity.Critical, "public class Two {\n    private IDependency _dependency;\n    \n    public One(IDependency dependency) {\n        _dependency = dependency;\n    }"),
+            // new(1, "Violation name 1", ViolationType.ForbiddenDependencyDirection, "A class from the 'Models' namespace must not depend on a class in the 'Controllers' namespace", ViolationSeverity.Major, "using Controllers;\n\nnamespace Models;\n\npublic class One {\n    private IDependency _dependency;\n    \n    public One(IDependency dependency) {\n        _dependency = dependency;\n    }"),
+            // new(2, "Violation name 2", ViolationType.Unknown, "Some other thing I have no idea...", ViolationSeverity.Minor, "public class Two {\n    private IDependency _dependency;\n    \n    public One(IDependency dependency) {\n        _dependency = dependency;\n    }"),
+            // new(3, "Violation name 3", ViolationType.Unknown, "Some other thing I have no idea...", ViolationSeverity.Critical, "public class Two {\n    private IDependency _dependency;\n    \n    public One(IDependency dependency) {\n        _dependency = dependency;\n    }"),
         };
     }
 }
